Create serialization utils package
