+++ app/src/main/java/nl/marc_apps/ovgo/data/api/dutch_railways/models/DutchRailwaysDeparture.kt
@@ -1,7 +1,7 @@
 package nl.marc_apps.ovgo.data.api.dutch_railways.models
 
 import kotlinx.serialization.Serializable
-import nl.marc_apps.ovgo.utils.DateSerializer
+import nl.marc_apps.ovgo.utils.serialization.DateSerializer
 import java.util.*
 
 @Serializable
