+++ app/src/main/java/nl/marc_apps/ovgo/data/api/dutch_railways/models/DutchRailwaysStation.kt
@@ -2,7 +2,7 @@ package nl.marc_apps.ovgo.data.api.dutch_railways.models
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
-import nl.marc_apps.ovgo.utils.DateSerializer
+import nl.marc_apps.ovgo.utils.serialization.DateSerializer
 import java.util.*
 
 @Serializable
